==> 1 Construccion de la red.Rmd <==
---
title: "Construcción de la red"
output: html_notebook
---

Se construye una red de oncogenes a partir del material proporcionado en el artículo en forma de archivo Excel.

En primer lugar cargamos (instalamos previamente si se requiere) las librerías necesarias. 
```{r}
#install.packages("igraph")
#install.packages("readxl")
library(igraph)
library(readxl)
```
Ahora leemos los datos a partir del documento Excel y construimos el grafo.
```{r}
aristas <- read_excel("./Dataset1.xlsx", 
                    sheet = "WeigthedNetwork")
nodos <- read_excel("./Dataset1.xlsx", 
                    sheet = "NetworkNodes")
grafo <- 
  graph_from_data_frame(d = aristas,  directed=F, vertices = nodos)
```

Estudiamos y visualizamos el grafo.
Se trata de una estructura con 177 nodos y 15576 aristas.
El grafo es ponderado y no dirigido. 
```{r}
V(grafo)
E(grafo)
class(grafo)
head(grafo)
```
```{r}
plot(grafo)
```

Probamos varias representaciones gráficas del grafo para intentar verlo con más claridad:

- Círculo
```{r}
l <- layout_in_circle(grafo)
plot(grafo, layout=l)

```
- Esfera
```{r}
e <- layout_on_sphere(grafo)
plot(grafo, layout=e)
```
- Fruchterman-Reingold
```{r}
fr <- layout_with_fr(grafo)
plot(grafo, layout=fr)
```
Las visualización del grafo es poco clara debido a la gran cantidad de aristas.

A continuación, lo representamos como mapa de calor a partir de la matriz de adyacencia.
El color más oscuro denota una relación más fuerte (peso de las aritas mayor) entre cada par de nodos.
```{r}
netm <- get.adjacency(grafo, attr="PESO", sparse=F)
colnames(netm) <- nodos$NODOS
rownames(netm) <- nodos$NODOS
palf <- colorRampPalette(c("gold", "dark orange")) 

heatmap(netm[,177:1], Rowv = NA, Colv = NA, col = palf(100), 
        scale="none", margins=c(3,1) )

```

Para entender mejor la red, representamos las aristas en función de su peso. Comprobamos que la mayoría tienen un peso inferior a 0.5.
La frecuencia de las aristas es mayor con valores de peso pequeños, y va disminuyendo a medida que aumenta el valor del peso. 
```{r}
hist(aristas$PESO, xlab = "Peso", ylab = "Frecuencia")
```
Se calculan también:
- Media del peso de las aristas: 0.2444883
- Desvación típica: 0.1691008
```{r}
mean(aristas$PESO)
sd(aristas$PESO)
```

Tras esta primera observacion del grafo, pensamos que su gran tamaño puede dificultar las tareas de clustering y aprendizaje supervisado, por lo que resolvemos podar el grafo a fin de facilitar estas operaciones.

Como paso previo a la poda, representamos el número de aristas en función de la densidad de estas y del número de subredes encontradas. 
La poda se realiza en función del peso de las aristas. Así, el umbral de poda está comprendido entre 0 y 1, incrementándose en 0.05 tras cada iteración. 
```{r}
umbral = seq(-0.1 , 1, by=0.05)
no = c(0)
ar = c(0)
i=1
for (x in umbral){
  gf <- delete_edges(grafo, E(grafo)[PESO<x])
  no[i] <- clusters(gf)$no #vector con el numero de clusters
  ar[i] <- edge_density(gf)
  i= i+1
}

plot (x= umbral, y =no, xlab = "Umbral" ,ylab ="Numero de subredes")
plot (x= umbral, y =ar, xlab = "Umbral" ,ylab ="Densidad de aristas")
```
Visualizando las gráficas, podemos ver que a patir de 0.4 el número de subredes crece exponencialmente. Para este valor, la denssiad de aristas ha disminuido notablemente, pero aún no se acerca demasiado a 0.
Por ello, tomamos 0.4 como el umbral de poda de las aristas del grafo. El grafo podado es sobre el que realizaremos las operaciones siguientes. 
```{r}
g4 <- delete_edges(grafo, E(grafo)[PESO<0.4])
```
Tomamos un contacto inicial con este nuevo grafo.
Tiene el mismo numero de vertices que el grafo de partida, 177.
Sin embargo, el numero de aristas se ha reducido hasta 3007 (un 19% de las aristas del grafo de partida).
Viendo la representacion gráfica observamos una mayor claridad. 
```{r}
V(g4)
E(g4)
plot(g4)
```
Observamos de nuevo la matriz de adyacencia.
Cuenta con muchas mas zonas de color amarillo claro, que denota una menor relación entre los nodos presentes.
```{r}
netm <- get.adjacency(g4, attr="PESO", sparse=F)
colnames(netm) <- nodos$NODOS
rownames(netm) <- nodos$NODOS
palf <- colorRampPalette(c("gold", "dark orange")) 

heatmap(netm[,177:1], Rowv = NA, Colv = NA, col = palf(100), 
        scale="none", margins=c(3,1) )
```
Visualizamos la matriz de distancias
```{r}
distM <- shortest.paths(g4, weights = E(g4)$PESO)
distM
```

Con esto hemos construido la red de oncogenes objeto de estudio y nos hemos acercado un poco a su topología.
