==> 3 Descubrimiento de comunidades.Rmd <==
---
title: "Descubrimiento de comunidades"
output: html_notebook
---

En el ámbito del analisis de redes biológicas, el descubrimiento de comunidades permite agrupar los nodos aludiendo a funcionalidad común o alto grado de relación.
En este Notebook se realiza un proceso de descubrimiento de comunidades para la red de oncogenes objeto de estudio.


En primer lugar, se cargan las bibliotecas necesarias.
```{r}
library(igraph)
library(readxl)
library(dplyr)
```

Se carga el grafo
```{r}
#Cargo datos
aristas <- read_excel("./Dataset1.xlsx", 
                    sheet = "WeigthedNetwork")
nodos <- read_excel("./Dataset1.xlsx", 
                    sheet = "NetworkNodes")
#Construyo el grafo
grafo_origen <- 
  graph_from_data_frame(d = aristas, vertices = nodos,  directed=F)

#Podo cogiendo como umbral 0.4
grafo <- delete_edges(grafo_origen, E(grafo_origen)[PESO<0.4])
```

Para el proceso de descubrimiento de comunidades se hará uso de dos algoritmos, el primero de ellos basado en propagación de etiquetas y el segundo, en optimización de la modularidad (ver Capítulo II).

- Detección de comunidades basada en propagación de etiquetas.

Como paso previo se genera un vector con números aleatorios que sirve para inicializar el algoritmo (si no, en cada ejecución dará resultados diferentes).
```{r}
set.seed(55)
inicio<-sample(c(0:176), size = 177, replace = TRUE)
```

Aplico el algoritmo y represento gráficamente el resultado. 
```{r}
clp <- cluster_label_prop(grafo, initial = inicio, weights = E(grafo))
plot(clp, grafo)
```
Observamos propiedades de las comunidades descubiertas.

Se estudian las comunidades y el número de nodos que tiene cada una:
Aparecen 19 comunidades:
- 18 con un solo nodo
- Comunidad 1: 159 nodos
```{r}
table(clp$membership)
```

A que comunidad (que etiqueta tiene) pertenece cada uno de los nodos.
Se pueden ver los nodos de las comunidades más numerosas. 
```{r}
comunidad <- clp$membership
miembros <- data.frame(nodos, comunidad)

c1 <- filter(miembros, comunidad == 1)
c1
```

Selecciono la comunidad con mayor numero de nodos (la 1) y la estudio en relación a las medidas de centralidad calculadas anteriormente.

- Degree (grado)
```{r}
deg <- degree(grafo, mode = "all")
degdf <- data.frame(nodos, deg)
juntos <-merge(x=degdf, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c1[order(-c1$deg),]
```

- Centrality (centralidad)
```{r}
cent_eing <- centr_eigen(grafo, directed = F, normalized = F)
eigencen <- data.frame(nodos, cent_eing$vector)
juntos <-merge(x=eigencen, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c1[order(-c1$cent_eing.vector),]
```

- Clossenes (cercanía)
```{r}
cent_clo <-centr_clo(grafo, mode = "all", normalized = F)
clossenes <- data.frame(nodos, cent_clo$res)
juntos <-merge(x=clossenes, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c1[order(-c1$cent_clo.res),]
```
- Betweenness (intermediación)
```{r}
bet <- betweenness(grafo, directed = F, weights = E(grafo))
betdf <- data.frame(nodos, bet)
juntos <-merge(x=betdf, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c1[order(-c1$bet),]
```
Los resultados se dan ordenados en base a los valores de los nodos para cada una de las medidas, separándolos por comunidades.


- Detección de comunidades basadas en optimización de modularidad.
```{r}
cfg <- cluster_fast_greedy(grafo, weights = E(grafo))
plot(cfg, grafo)
```
Observamos propiedades de las comunidades descubiertas.

Se estudian las comunidades y el número de nodos que tiene cada una.
Aparecen 22 comunidades:
- 18 con un solo nodo
- Comunidad 1: 17 nodos
- Comunidad 2: 68 nodos
- Comunidad 3: 72 nodos
- Comunidad 4: 2 nodos
```{r}
table(cfg$membership)
```

Comunidad a la que pertenece cada nodo.
Se imprimen aquellas con más de un nodo. 
```{r}
comunidad <- cfg$membership
miembros <- data.frame(nodos, comunidad)

c4 <- filter(miembros, comunidad == 4)
c4
c3 <- filter(miembros, comunidad == 3)
c3
c2 <- filter(miembros, comunidad == 2)
c2
c1 <- filter(miembros, comunidad == 1)
c1

```

Selecciono las comunidades con mayor numero de nodos (la 1, la 2 y la 3) y las estudio en relación a las medidas de centralidad calculadas anteriormente.

- Degree (grado)
```{r}
deg <- degree(grafo, mode = "all")
degdf <- data.frame(nodos, deg)
juntos <-merge(x=degdf, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c2 <- filter(juntos, comunidad == 2)
c3 <- filter(juntos, comunidad == 3)
c1[order(-c1$deg),]
c2[order(-c2$deg),]
c3[order(-c3$deg),]
```

- Centrality (centralidad)
```{r}
cent_eing <- centr_eigen(grafo, directed = F, normalized = T)
eigencen <- data.frame(nodos, cent_eing$vector)
juntos <-merge(x=eigencen, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c2 <- filter(juntos, comunidad == 2)
c3 <- filter(juntos, comunidad == 3)
c1[order(-c1$cent_eing.vector),]
c2[order(-c2$cent_eing.vector),]
c3[order(-c3$cent_eing.vector),]
```

- Clossenes (cercanía)
```{r}
cent_clo <-centr_clo(grafo, mode = "all", normalized = T)
clossenes <- data.frame(nodos, cent_clo$res)
juntos <-merge(x=clossenes, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c2 <- filter(juntos, comunidad == 2)
c3 <- filter(juntos, comunidad == 3)
c1[order(-c1$cent_clo.res),]
c2[order(-c2$cent_clo.res),]
c3[order(-c3$cent_clo.res),]
```

- Betweenness (intermediación)
```{r}
bet <- betweenness(grafo, directed = F, weights = E(grafo))
betdf <- data.frame(nodos, bet)
juntos <-merge(x=betdf, miembros, by= "NODOS")
c1 <- filter(juntos, comunidad == 1)
c2 <- filter(juntos, comunidad == 2)
c3 <- filter(juntos, comunidad == 3)
c1[order(-c1$bet),]
c2[order(-c2$bet),]
c3[order(-c3$bet),]
```
Los resultados se dan ordenados en base a los valores de los nodos para cada una de las medidas, separándolos por comunidades.



Guardo el listado de miembros de la comunidad en un archivo txt que será usado posteriormente.
```{r}
write.csv(miembros, file="comunidades.txt")
```
