==> 4 Generacion dataset.Rmd <==
---
title: "Generación del dataset"
output: html_notebook
---

En este Notebook se realiza la construcción de un dataset formado por el listado de nodos y el valor de las medidas de centralidad para ellos.  

En primer lugar se cargan las librerias necesarias
```{r}
library(igraph)
library(readxl)
```

Cargamos los datos del grafo.
```{r}
#Cargo datos
aristas <- read_excel("./Dataset1.xlsx", 
                    sheet = "WeigthedNetwork")
nodos <- read_excel("./Dataset1.xlsx", 
                    sheet = "NetworkNodes")
#Construyo el grafo
grafo_origen <- 
  graph_from_data_frame(d = aristas, vertices = nodos,  directed=F)

#Podo cogiendo como umbral 0.4
grafo <- delete_edges(grafo_origen, E(grafo_origen)[PESO<0.4])
```

Calculo los vectores con las medidas que quiero que tenga el dataframe (degree, centrality, closseness, betweenness).
```{r}
#Degree
deg <- degree(grafo, mode = "all")

#Closeness
close <- closeness(grafo, mode= "all", normalized = T)

#Centrality
eig_cent <- eigen_centrality(grafo, directed = F)

#Betweenness
bet <- betweenness(grafo, directed = F, weights = E(grafo))

```

Este dataset será usado para aprendizaje supervisado.
Por ello, debe contener un campo que funcione como etiqueta cuyo valor intentaremos predecir.
Habrá dos campos etiqueta, uno basado en la información obtenida en el descubrimiento de comunidades ("ser nodo aislado") y otro en conocimiento experto, "ser oncogen de la lista de NCG".

- Ser (1) o no (0) nodo aislado, en referencia a los datos obtenidos con el descubrimiento de comunidades.

Para ello, primero leo el listado de nodos con las comunidades a las que pertenecen.
```{r}
df <- read.table("./comunidades.txt", header = TRUE, sep = ",")
```
En este listado, los nodos aislados son los que pertenecen a comunidades distintas a la 1, 2 y 3. 
```{r}
l1 <- c(0)
indices = seq(1 , 177, by=1)
for(i in indices){
  n <- 0
  if(df$comunidad[i] > 3){
    n <- 1
  }
  l1[i] <- n
}
```

- Ser (1) o no (0) oncogen de la red de NCG, en base a conocimiento experto

Para ello, cargamos el listado de genes
```{r}
genes <- read.delim('cancergenes_list.txt', header = TRUE, sep = "\t")
```
Comparamos la lista de oncogenes de NCG con los nodos del grafo. 
```{r}
colnames(genes) <- c("NODOS", "candidatos")
en_comun <- merge(nodos, genes)
known <- en_comun$NODOS
```
Comparamos la lista de genes candidatos a ser oncogenes de NCG con los nodos del grafo.
```{r}
colnames(genes) <- c("Oncogenes", "NODOS")
en_comun <- merge(nodos, genes)
candidate <- en_comun$NODOS
```
Los unifico todos en una lista
```{r}
lista <- c(known, candidate)
```
Unifico con los nodos para crear una lista de 1 y 0 donde solo tenga un 1 los que pertenezcan a esta lista. 
```{r}
l2 <- c(0)
indices = seq(1 , 177, by=1)
for(i in indices){
  n <- 0
  for(y in lista){
    if(nodos$NODOS[i] == y)
      n <- 1
  }
  l2[i] <- n
}
```

Genero el dataset creando un dataframe con los vectores obtenidos anteriormente.
```{r}
grafodf <- data.frame(
  "Degree" = deg,
  "Centrality" = eig_cent$vector,
  "Closeness" = close, 
  "Betweenness" = bet,
  "Nodo aislado" = l1,
  "Estar NCG" = l2
)
```

Finalmente el dataframe se guarda como archivo de texto para poder ser importado posteriormente. 
```{r}
write.csv(grafodf, file="dataframe.txt")
```
